+++ test/test_decode.cpp
@@ -4,31 +4,68 @@ using namespace mettle;
 #include "bencode.hpp"
 
 template<typename T>
-auto make_data(const char *data) ->
-std::enable_if_t<std::is_constructible_v<T, const char *>, T> {
-  return T((char*)(data));
-}
+struct char_type {
+  using type = typename T::value_type;
+};
+
+template<typename Char>
+struct char_type<Char *> { using type = std::remove_cv_t<Char>; };
+template<typename Char>
+struct char_type<std::basic_istringstream<Char>> { using type = Char; };
 
 template<typename T>
-auto make_data(const char *data) ->
-std::enable_if_t<!std::is_constructible_v<T, const char *>, T> {
-  return T(data, data + std::strlen(data));
+inline constexpr bool supports_view =
+  !std::is_same_v<typename char_type<T>::type, std::byte>;
+template<typename Char>
+inline constexpr bool supports_view<std::basic_istringstream<Char>> = false;
+
+template<typename T>
+T make_data(const char *data) {
+  using Char = typename char_type<T>::type;
+  if constexpr(std::is_constructible_v<T, const Char *>)
+    return T((Char*)(data));
+  else
+    return T((Char*)data, (Char*)(data + std::strlen(data)));
 }
 
-struct at_eof : matcher_tag {
-  bool operator ()(const char *s) const {
-    return *s == '\0';
-  }
-
-  template<typename Char, typename Traits>
-  bool operator ()(const std::basic_ios<Char, Traits> &ss) const {
-    return ss.eof();
-  }
-
+struct at_eof_base : matcher_tag {
   std::string desc() const {
     return "at eof";
   }
 };
+
+template<typename ...T>
+struct at_eof;
+
+template<typename Char>
+struct at_eof<Char *> : at_eof_base {
+  at_eof(const Char *) {}
+  bool operator ()(const Char *s) const { return *s == Char('\0'); }
+};
+
+template<typename Char>
+struct at_eof<Char *, std::size_t> : at_eof_base {
+  at_eof(const Char *start, std::size_t len) : end(start + len) {}
+  bool operator ()(const Char *s) const { return s == end; }
+  Char * end;
+};
+
+template<typename Iter>
+struct at_eof<Iter, Iter> : at_eof_base {
+  at_eof(Iter, Iter end) : end(end) {}
+  bool operator ()(Iter i) const { return i == end; }
+  Iter end;
+};
+
+template<typename Char>
+struct at_eof<std::basic_istringstream<Char>> : at_eof_base {
+  using ios_type = std::basic_istringstream<Char>;
+  at_eof(ios_type&) {}
+  bool operator ()(const ios_type &ss) const { return ss.eof(); }
+};
+
+template<typename ...T>
+at_eof(T &&...t) -> at_eof<std::remove_reference_t<std::remove_cv_t<T>>...>;
 
 template<typename T>
 auto within_memory(const T &) {
@@ -42,7 +79,7 @@ auto within_memory(const T &t) {
 
 template<typename T>
 auto within_memory(const T *t) {
-  return in_interval(t, t + std::strlen(t), interval::closed);
+  return in_interval(t, t + bencode::detail::any_strlen(t), interval::closed);
 }
 
 template<typename Nested, typename Matcher>
@@ -71,16 +108,26 @@ auto decode_error(const std::string &what, std::size_t offset) {
   );
 }
 
+template<typename DecodeArgs, typename DoDecode>
+auto mix_decoder(DecodeArgs &&decode_args, DoDecode &&do_decode) {
+  return [decode_args, do_decode](auto &&data) {
+    return std::apply(
+      do_decode, decode_args(std::forward<decltype(data)>(data))
+    );
+  };
+}
+
 template<typename InType, typename Builder, typename Callable>
 auto decode_tests(Builder &_, Callable &&decode) {
   using boost::get;
   using std::get;
 
-  using OutType = fixture_type_t<decltype(_)>;
+  using OutType = decltype(decode(std::declval<InType>()));
   using Integer = typename OutType::integer;
   using String = typename OutType::string;
   using List = typename OutType::list;
   using Dict = typename OutType::dict;
+  auto S = &make_data<String>;
 
   _.test("integer", [decode]() {
     auto pos = make_data<InType>("i42e");
@@ -92,14 +139,14 @@ auto decode_tests(Builder &_, Callable &&decode) {
     expect(get<Integer>(neg_value), equal_to(-42));
   });
 
-  _.test("string", [decode]() {
+  _.test("string", [decode, S]() {
     auto data = make_data<InType>("4:spam");
     // Silence GCC < 10.
     [[maybe_unused]] auto within_data_memory = within_memory(data);
 
     auto value = decode(data);
     auto str = get<String>(value);
-    expect(str, equal_to("spam"));
+    expect(str, equal_to(S("spam")));
     if constexpr(bencode::detail::is_view<String>) {
       expect(&*str.begin(), within_data_memory);
       expect(&*str.end(), within_data_memory);
@@ -113,24 +160,24 @@ auto decode_tests(Builder &_, Callable &&decode) {
     expect(get<Integer>(list[0]), equal_to(42));
   });
 
-  _.test("dict", [decode]() {
+  _.test("dict", [decode, S]() {
     auto data = make_data<InType>("d4:spami42ee");
     // Silence GCC < 10.
     [[maybe_unused]] auto within_data_memory = within_memory(data);
 
     auto value = decode(data);
     auto dict = get<Dict>(value);
-    expect(get<Integer>(dict["spam"]), equal_to(42));
+    expect(get<Integer>(dict[S("spam")]), equal_to(42));
 
-    auto str = dict.find("spam")->first;
-    expect(str, equal_to("spam"));
+    auto str = dict.find(S("spam"))->first;
+    expect(str, equal_to(S("spam")));
     if constexpr(bencode::detail::is_view<String>) {
       expect(&*str.begin(), within_data_memory);
       expect(&*str.end(), within_data_memory);
     }
   });
 
-  _.test("nested", [decode]() {
+  _.test("nested", [decode, S]() {
     auto data = make_data<InType>(
       "d"
       "3:one" "i1e"
@@ -140,161 +187,316 @@ auto decode_tests(Builder &_, Callable &&decode) {
     );
     auto value = decode(data);
     auto dict = get<Dict>(value);
-    expect(get<Integer>(dict["one"]), equal_to(1));
-    expect(get<String>(get<List>(dict["two"])[1]), equal_to("foo"));
-    expect(get<Integer>(get<Dict>(get<List>(dict["three"])[0])["foo"]),
+    expect(get<Integer>(dict[S("one")]), equal_to(1));
+    expect(get<String>(get<List>(dict[S("two")])[1]), equal_to(S("foo")));
+    expect(get<Integer>(get<Dict>(get<List>(dict[S("three")])[0])[S("foo")]),
            equal_to(0));
   });
 }
 
-suite<> test_decode("test decoder", [](auto &_) {
-  subsuite<
-    const char *, std::string, std::vector<char>, std::istringstream
-  >(_, "decode", type_only, [](auto &_) {
-    using InType = fixture_type_t<decltype(_)>;
+template<typename DecodeArgs>
+struct decode_suites {
+  decode_suites(DecodeArgs decode_args)
+    : decode_args(std::move(decode_args)) {}
+  DecodeArgs decode_args;
 
-    subsuite<
-      bencode::data, bencode::boost_data
-    >(_, "decode to", type_only, [](auto &_) {
-      using OutType = fixture_type_t<decltype(_)>;
-      decode_tests<InType>(_, [](auto &&data) {
-        auto value = bencode::basic_decode<OutType>(data);
-        if constexpr(std::is_same_v<InType, std::istringstream>) {
-          expect(data, at_eof());
-        }
+  template<typename Builder>
+  void operator ()(Builder &_) const {
+    using InType = fixture_type_t<decltype(_)>;
+    using Char = typename char_type<InType>::type;
+    auto &decode_args = this->decode_args;
+
+    subsuite<>(_, "decode", [decode_args](auto &_) {
+      decode_tests<InType>(_, mix_decoder(decode_args,
+                                          [](auto &&data, auto &&...rest) {
+        auto value = bencode::decode(data, rest...);
+        if constexpr(std::is_base_of_v<std::ios_base, InType>)
+          expect(data, at_eof(data));
         return value;
-      });
+      }));
     });
 
-    if constexpr(!std::is_same_v<InType, std::istringstream>) {
-      subsuite<
-        bencode::data_view, bencode::boost_data_view
-      >(_, "decode to", type_only, [](auto &_) {
-        using OutType = fixture_type_t<decltype(_)>;
-        decode_tests<InType>(_, [](auto &&data) {
-          return bencode::basic_decode<OutType>(data);
-        });
+    if constexpr(supports_view<InType>) {
+      subsuite<>(_, "decode_view", [decode_args](auto &_) {
+        decode_tests<InType>(_, mix_decoder(decode_args,
+                                            [](auto &&data, auto &&...rest) {
+          return bencode::decode_view(data, rest...);
+        }));
       });
     }
-  });
 
-  subsuite<
-    std::string, std::vector<char>
-  >(_, "decode iterator pair", type_only, [](auto &_) {
-    using InType = fixture_type_t<decltype(_)>;
-
-    subsuite<
-      bencode::data, bencode::boost_data, bencode::data_view,
-      bencode::boost_data_view
-    >(_, "decode to", type_only, [](auto &_) {
-      using OutType = fixture_type_t<decltype(_)>;
-      decode_tests<InType>(_, [](auto &&data) {
-        return bencode::basic_decode<OutType>(data.begin(), data.end());
-      });
-    });
-  });
-
-  subsuite<>(_, "decode pointer/length", [](auto &_) {
-    subsuite<
-      bencode::data, bencode::boost_data, bencode::data_view,
-      bencode::boost_data_view
-    >(_, "decode to", type_only, [](auto &_) {
-      using OutType = fixture_type_t<decltype(_)>;
-      decode_tests<const char *>(_, [](const char *data) {
-        return bencode::basic_decode<OutType>(data, data + std::strlen(data));
-      });
-    });
-  });
-
-  subsuite<
-    const char *, std::istringstream
-  >(_, "decode_some", type_only, [](auto &_) {
-    using InType = fixture_type_t<decltype(_)>;
-
-    subsuite<
-      bencode::data, bencode::boost_data
-    >(_, "decode to", type_only, [](auto &_) {
-      using OutType = fixture_type_t<decltype(_)>;
-      decode_tests<InType>(_, [](auto &&data) {
-        auto value = bencode::basic_decode_some<OutType>(data);
-        expect(data, at_eof());
+    subsuite<>(_, "boost_decode", [decode_args](auto &_) {
+      decode_tests<InType>(_, mix_decoder(decode_args,
+                                          [](auto &&data, auto &&...rest) {
+        auto value = bencode::boost_decode(data, rest...);
+        if constexpr(std::is_base_of_v<std::ios_base, InType>)
+          expect(data, at_eof(data));
         return value;
-      });
-
-      _.test("successive objects", []() {
-        auto data = make_data<InType>("i42e4:goat");
-
-        auto first = bencode::decode_some(data);
-        expect(std::get<bencode::integer>(first), equal_to(42));
-        expect(data, is_not(at_eof()));
-
-        auto second = bencode::decode_some(data);
-        expect(std::get<bencode::string>(second), equal_to("goat"));
-        expect(data, at_eof());
-      });
+      }));
     });
 
-    if constexpr(!std::is_same_v<InType, std::istringstream>) {
+    if constexpr(supports_view<InType>) {
+      subsuite<>(_, "boost_decode_view", [decode_args](auto &_) {
+        decode_tests<InType>(_, mix_decoder(decode_args,
+                                            [](auto &&data, auto &&...rest) {
+          return bencode::boost_decode_view(data, rest...);
+        }));
+      });
+    }
+
+    subsuite<
+      bencode::data_for_char_t<Char>, bencode::boost_data_for_char_t<Char>
+    >(_, "basic_decode to", type_only, [decode_args](auto &_) {
+      using OutType = fixture_type_t<decltype(_)>;
+      decode_tests<InType>(_, mix_decoder(decode_args,
+                                          [](auto &&data, auto &&...rest) {
+        auto value = bencode::basic_decode<OutType>(data, rest...);
+        if constexpr(std::is_base_of_v<std::ios_base, InType>)
+          expect(data, at_eof(data));
+        return value;
+      }));
+    });
+
+    if constexpr(supports_view<InType>) {
       subsuite<
-        bencode::data_view, bencode::boost_data_view
-      >(_, "decode to", type_only, [](auto &_) {
+        bencode::data_view_for_char_t<Char>,
+        bencode::boost_data_view_for_char_t<Char>
+      >(_, "basic_decode to", type_only, [decode_args](auto &_) {
         using OutType = fixture_type_t<decltype(_)>;
-        decode_tests<InType>(_, [](auto &&data) {
-          auto value = bencode::basic_decode_some<OutType>(data);
-          expect(data, at_eof());
+        decode_tests<InType>(_, mix_decoder(decode_args,
+                                            [](auto &&data, auto &&...rest) {
+          auto value = bencode::basic_decode<OutType>(data, rest...);
+          if constexpr(std::is_base_of_v<std::ios_base, InType>)
+            expect(data, at_eof(data));
           return value;
-        });
+        }));
       });
     }
+  }
+};
+
+template<typename InType, typename Builder, typename DecodeArgs,
+         typename DoDecode>
+auto decode_some_tests(Builder &_, DecodeArgs &&decode_args,
+                       DoDecode &&do_decode) {
+  using boost::get;
+  using std::get;
+
+  _.test("successive objects", [decode_args, do_decode]() {
+    auto data = make_data<InType>("i42e4:goat");
+    auto args = decode_args(data);
+    auto at_end = std::apply([](auto &&...t) {
+      return at_eof(std::forward<decltype(t)>(t)...);
+    }, args);
+
+    auto first = std::apply(do_decode, args);
+    using OutType = decltype(first);
+    using Integer = typename OutType::integer;
+    using String = typename OutType::string;
+    auto S = &make_data<String>;
+
+    expect(get<Integer>(first), equal_to(42));
+    expect(std::get<0>(args), is_not(at_end));
+
+    auto second = std::apply(do_decode, args);
+    expect(get<String>(second), equal_to(S("goat")));
+    expect(std::get<0>(args), at_end);
   });
+}
+
+template<typename DecodeArgs>
+struct decode_some_suites {
+  decode_some_suites(DecodeArgs decode_args)
+    : decode_args(std::move(decode_args)) {}
+  DecodeArgs decode_args;
+
+  template<typename Builder>
+  void operator ()(Builder &_) const {
+    using InType = fixture_type_t<decltype(_)>;
+    using Char = typename char_type<InType>::type;
+    auto &decode_args = this->decode_args;
+
+    subsuite<>(_, "decode_some", [decode_args](auto &_) {
+      decode_tests<InType>(
+        _, mix_decoder(decode_args, [](auto &&data, auto &&...rest) {
+          auto at_end = at_eof(data, rest...);
+          auto value = bencode::decode_some(data, rest...);
+          expect(data, at_end);
+          return value;
+        })
+      );
+      decode_some_tests<InType>(
+        _, decode_args, [](auto &&data, auto &&...rest) {
+          return bencode::decode_some(data, rest...);
+        }
+      );
+    });
+
+    if constexpr(supports_view<InType>) {
+      subsuite<>(_, "decode_view_some", [decode_args](auto &_) {
+        decode_tests<InType>(
+          _, mix_decoder(decode_args, [](auto &&data, auto &&...rest) {
+            auto at_end = at_eof(data, rest...);
+            auto value = bencode::decode_view_some(data, rest...);
+            expect(data, at_end);
+            return value;
+          })
+        );
+        decode_some_tests<InType>(
+          _, decode_args, [](auto &&data, auto &&...rest) {
+            return bencode::decode_view_some(data, rest...);
+          }
+        );
+      });
+    }
+
+    subsuite<>(_, "boost_decode_some", [decode_args](auto &_) {
+      decode_tests<InType>(
+        _, mix_decoder(decode_args, [](auto &&data, auto &&...rest) {
+          auto at_end = at_eof(data, rest...);
+          auto value = bencode::boost_decode_some(data, rest...);
+          expect(data, at_end);
+          return value;
+        })
+      );
+      decode_some_tests<InType>(
+        _, decode_args, [](auto &&data, auto &&...rest) {
+          return bencode::boost_decode_some(data, rest...);
+        }
+      );
+    });
+
+    if constexpr(supports_view<InType>) {
+      subsuite<>(_, "boost_decode_view_some", [decode_args](auto &_) {
+        decode_tests<InType>(
+          _, mix_decoder(decode_args, [](auto &&data, auto &&...rest) {
+            auto at_end = at_eof(data, rest...);
+            auto value = bencode::boost_decode_view_some(data, rest...);
+            expect(data, at_end);
+            return value;
+          })
+        );
+        decode_some_tests<InType>(
+          _, decode_args, [](auto &&data, auto &&...rest) {
+            return bencode::boost_decode_view_some(data, rest...);
+          }
+        );
+      });
+    }
+
+    subsuite<
+      bencode::data_for_char_t<Char>, bencode::boost_data_for_char_t<Char>
+    >(_, "basic_decode_some to", type_only, [decode_args](auto &_) {
+      using OutType = fixture_type_t<decltype(_)>;
+      decode_tests<InType>(
+        _, mix_decoder(decode_args, [](auto &&data, auto &&...rest) {
+          auto at_end = at_eof(data, rest...);
+          auto value = bencode::basic_decode_some<OutType>(data, rest...);
+          expect(data, at_end);
+          return value;
+        })
+      );
+      decode_some_tests<InType>(
+        _, decode_args, [](auto &&data, auto &&...rest) {
+          return bencode::basic_decode_some<OutType>(data, rest...);
+        }
+      );
+    });
+
+    if constexpr(supports_view<InType>) {
+      subsuite<
+        bencode::data_view_for_char_t<Char>,
+        bencode::boost_data_view_for_char_t<Char>
+      >(_, "basic_decode_some to", type_only, [decode_args](auto &_) {
+        using OutType = fixture_type_t<decltype(_)>;
+        decode_tests<InType>(
+          _, mix_decoder(decode_args, [](auto &&data, auto &&...rest) {
+            auto at_end = at_eof(data, rest...);
+            auto value = bencode::basic_decode_some<OutType>(data, rest...);
+            expect(data, at_end);
+            return value;
+          })
+        );
+        decode_some_tests<InType>(
+          _, decode_args, [](auto &&data, auto &&...rest) {
+            return bencode::basic_decode_some<OutType>(data, rest...);
+          }
+        );
+      });
+    }
+  }
+};
+
+suite<> test_decode("test decoder", [](auto &_) {
+  using std::istringstream;
+#if __cpp_char8_t
+  using u8istringstream = std::basic_istringstream<char8_t>;
+#endif
+
+  auto decode_args = [](auto &&data) {
+    return std::tuple<decltype(data)&>(data);
+  };
+  auto decode_iter_args = [](auto &&data) {
+    return std::make_tuple(data.begin(), data.end());
+  };
+  auto decode_ptr_len_args = [](auto &&data) {
+    return std::tuple<decltype(data)&, std::size_t>(
+      data, bencode::detail::any_strlen(data)
+    );
+  };
 
   subsuite<
-    std::string, std::vector<char>
-  >(_, "decode_some iterator pair", type_only, [](auto &_) {
-    using InType = fixture_type_t<decltype(_)>;
+    const char *,      std::string,   std::vector<char>,      istringstream,
+#if __cpp_char8_t
+    const char8_t *,   std::u8string, std::vector<char8_t>,   u8istringstream,
+#endif
+    const std::byte *,                std::vector<std::byte>
+  >(_, "decode", type_only, decode_suites(decode_args));
 
-    subsuite<
-      bencode::data, bencode::boost_data, bencode::data_view,
-      bencode::boost_data_view
-    >(_, "decode to", type_only, [](auto &_) {
-      using OutType = fixture_type_t<decltype(_)>;
-      decode_tests<InType>(_, [](auto &&data) {
-        auto begin = data.begin(), end = data.end();
-        auto value = bencode::basic_decode_some<OutType>(begin, end);
-        expect(begin, equal_to(end));
-        return value;
-      });
+  subsuite<
+    std::string,   std::vector<char>,
+#if __cpp_char8_t
+    std::u8string, std::vector<char8_t>,
+#endif
+    std::vector<std::byte>
+  >(_, "decode iterator pair", type_only, decode_suites(decode_iter_args));
 
-      _.test("successive objects", []() {
-        auto data = make_data<InType>("i42e4:goat");
-        auto begin = data.begin(), end = data.end();
+  subsuite<
+    const char *,
+#if __cpp_char8_t
+    const char8_t *,
+#endif
+    const std::byte *
+  >(_, "decode pointer/length", type_only, decode_suites(decode_ptr_len_args));
 
-        auto first = bencode::decode_some(begin, end);
-        expect(std::get<bencode::integer>(first), equal_to(42));
-        expect(begin, less(end));
+  subsuite<
+    const char *,      istringstream,
+#if __cpp_char8_t
+    const char8_t *,   u8istringstream,
+#endif
+    const std::byte *
+  >(_, "decode_some", type_only, decode_some_suites(decode_args));
 
-        auto second = bencode::decode_some(begin, end);
-        expect(std::get<bencode::string>(second), equal_to("goat"));
-        expect(begin, equal_to(end));
-      });
-    });
-  });
+  subsuite<
+    std::string, std::vector<char>,
+#if __cpp_char8_t
+    std::u8string, std::vector<char8_t>,
+#endif
+    std::vector<std::byte>
+  >(_, "decode_some iterator pair", type_only,
+    decode_some_suites([](auto &&data) {
+      return std::make_tuple(data.begin(), data.end());
+    })
+  );
 
-  subsuite<>(_, "decode_some pointer/length", [](auto &_) {
-    subsuite<
-      bencode::data, bencode::boost_data, bencode::data_view,
-      bencode::boost_data_view
-    >(_, "decode to", type_only, [](auto &_) {
-      using OutType = fixture_type_t<decltype(_)>;
-      decode_tests<const char *>(_, [](const char *data) {
-        auto value = bencode::basic_decode_some<OutType>(
-          data, data + std::strlen(data)
-        );
-        expect(data, at_eof());
-        return value;
-      });
-    });
-  });
+  subsuite<
+    const char *,
+#if __cpp_char8_t
+    const char8_t *,
+#endif
+    const std::byte *
+  >(_, "decode_some pointer/length", type_only,
+    decode_some_suites(decode_ptr_len_args));
 
   subsuite<>(_, "decoding integers", [](auto &_) {
     using udata = bencode::basic_data<